# Efficient implementation of `Tensor::ones()` for `metal`