--- a/candle-core/tests/tensor_tests.rs
+++ b/candle-core/tests/tensor_tests.rs
@@ -29,6 +29,14 @@ fn ones(device: &Device) -> Result<()> {
         Tensor::ones((2, 3), DType::F64, device)?.to_vec2::<f64>()?,
         [[1.0, 1.0, 1.0], [1.0, 1.0, 1.0]],
     );
+    assert_eq!(
+        Tensor::ones((2, 3), DType::F16, device)?.to_vec2::<half::f16>()?,
+        [[half::f16::from_f32(1.0), half::f16::from_f32(1.0), half::f16::from_f32(1.0)], [half::f16::from_f32(1.0), half::f16::from_f32(1.0), half::f16::from_f32(1.0)]],
+    );
+    assert_eq!(
+        Tensor::ones((2, 3), DType::BF16, device)?.to_vec2::<half::bf16>()?,
+        [[half::bf16::from_f32(1.0), half::bf16::from_f32(1.0), half::bf16::from_f32(1.0)], [half::bf16::from_f32(1.0), half::bf16::from_f32(1.0), half::bf16::from_f32(1.0)]],
+    );
     Ok(())
 }