--- a/candle-metal-kernels/src/fill.metal
+++ b/candle-metal-kernels/src/fill.metal
@@ -5,8 +5,12 @@ using namespace metal;
 template<typename T> METAL_FUNC void fill_with(
     device T *out,
     constant float &value,
+    constant size_t &numel,
     uint tid [[thread_position_in_grid]]
 ) {
+    if (tid >= numel) {
+        return;
+    }
     out[tid] = static_cast<T>(value);
 }
 
@@ -14,9 +18,10 @@ template<typename T> METAL_FUNC void fill_with(
 kernel void fill_##NAME(                                \
     device T *out,                                      \
     constant float &value,                              \
+    constant size_t &numel,                              \
     uint tid [[thread_position_in_grid]]                \
 ) {                                                     \
-    fill_with<T>(out, value, tid);                      \
+    fill_with<T>(out, value, numel, tid);              \
 }                                                       \