@AnubhabB 2024-10-01
@LaurentMazare `rustfmt` is faicling .. should I run a `cargo fmt --all` and push changes?